move display related JS to components folder
