+++ pages/resources/index.tsx
@@ -3,8 +3,8 @@ import styles from "../../styles/Home.module.css";
 import Link from "next/link";
 import { useContext } from "react";
 import { AppContext } from "../../context/context";
-import { Resource } from "../../models/types";
 import { useResources } from "../../hooks/useResources";
+import { ResourcesDisplay } from "../../components/resources/ResourcesDisplay";
 
 const Resources: NextPage = () => {
   const quizState = useContext(AppContext)
@@ -12,14 +12,13 @@ const Resources: NextPage = () => {
 
   if (isError) return <div>failed to load</div>
   if (isLoading) return <div>loading...</div>
-  if (!resources) return <div>failed to load</div>
+  if (!resources) return <div>Internal server error: invalid resources loaded</div>
 
-  return (
+  return  (
+
     <div className={styles.container}>
       <h1>Results</h1>
-      <div>
-        {resources.map(resource => <div key={resource['id']}><Link href={`resources/${resource['id']}`} >{resource['name']}</Link><br /></div>)}
-      </div>
+      <ResourcesDisplay resources={resources}/>
       <Link href='/quiz'>Back to Quiz</Link>
     </div>
   )
